pyproject.toml
[tool.poetry]
name = "langgraph-template"
version = "0.1.0"
description = "A minimal, reusable LangGraph template with multi-agent workflows"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
langchain-tavily = "^0.1.0"
sqlalchemy = "^2.0.0"
chromadb = "^0.4.0"
langchain-chroma = "^0.1.0"
chainlit = "^2.9.0"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
